this introduces the following main changes: (1) doubles are changed to floats. (2) in the code for calculating eta derivatives, a dot product was replaced by norm_dot to fix a bug that caused all derivatives to be wrong. (3) previously, some interface-interface intersections, in which the perimeters barely touched, were discarded by GETAREA but counted by TRIFORCE; now, TRIFORCE uses a threshold to discard of them as well, realigning the derivatives in some outlier cases. (4) there could probably be some other changes due to further implementation of semi analytic sea water

+++ triforce/benchmark.cc
@@ -21,11 +21,11 @@ void Benchmark::start(string phase){
 }
 void Benchmark::stop(){
 	clock_t clock_end;
-	double t;
-	pair<string,double> p;
+	float t;
+	pair<string,float> p;
 	clock_end=ms();
 	
-	t=(double)(clock_end-clock_start);
+	t=(float)(clock_end-clock_start);
 	if(times.find(phase)==times.end()){
 		p.first=string(phase);
 		p.second=t;
@@ -51,8 +51,8 @@ void Benchmark::print(FILE* outputfile){
 }
 
 
-void Benchmark::addQuantity(string quantity, double x){
-	pair<string,double> p;
+void Benchmark::addQuantity(string quantity, float x){
+	pair<string,float> p;
 	if(stats.find(quantity)==stats.end()){
 		p.first=string(quantity);
 		p.second=x;
